Create separate directories for custom testing platform targets
