+++ packages/constraint-engine/src/customTargets/testingPlatform/packageA/buildTestingPlatformPackageAReference.test.ts
@@ -4,7 +4,7 @@ import {
   signaler,
 } from 'rat-test/type-script/transgressing';
 import { fileSystemUtil } from 'mouse-test/type-script/shell';
-import { TestingPlatformTargetTypeId } from './targets';
+import { TestingPlatformTargetTypeId } from '../targets';
 import { buildTestingPlatformPackageAReference } from './buildTestingPlatformPackageAReference';
 
 // eslint-disable-next-line @typescript-eslint/no-floating-promises
